Create CryptoDad.sol
This is not my own code but rather a project I followed along to from a youtube video. I found it very helpful and hands on to code with a video as he explained how different functions worked and why and I encountered errors on my own that made me want to pull my hair out but were so satisfying once i did solve them.